Tests/LibWeb/Text/expected/wheel-scroll-over-iframe-at-boundary-falls-back-to-parent.txt

iframe.scrollTop before wheel: 450
iframe.scrollHeight: 600
iframe.clientHeight: 150
window.scrollY: 100
iframe.scrollTop after wheel: 450
